Basic CSS imports added to webpack
This approach will take the imported css and add javascript to the bundle in order to inject a style tag containing the parsed css. This is maybe not optimal because the css is all contained within the javascript instead of being added to a separate bundle css, but it works.