Add html-webpack-plugin to autoinject script tags
Now that we have code splitting implemented, we need to update our index.html to add the script tags. This is bothersome, so we use the html-webpack-plugin to do it for us. We move our index.html into /src and use it as a template. Then, webpack gives us an index.html in /dist with the script tags added automagically.